Update MSLK in GB200 flagship feedstock recipe (#70)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/MSLK/pull/70

Update MSLK in GB200 flagship feedstock recipe.  This updates the build scripts to pull the versions of the dependencies pinned in the repo, so that we don't need to manually update it.

Note that even though MSLK is included in the feedstock, it is currently **not** used/adopted in any code, so this should be a no-op

Reviewed By: cthi

Differential Revision: D89487097

fbshipit-source-id: 3c8d359789256dcf9357d4cebdb02f5f4d8f3e7b